ai-call-assistant-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
twilio==9.0.4
anthropic==0.25.0
supabase==2.4.2
firebase-admin==6.4.0
python-dotenv==1.0.1
pydantic>=2.0.0
pydantic-core>=2.0.0
httpx==0.27.0
python-multipart==0.0.9
gunicorn==21.2.0
